+++ client-js/rtvi/events.ts
@@ -5,7 +5,7 @@
  */
 
 import { MediaState, Participant, TransportState } from "./common_types";
-import { DeviceError } from "./errors";
+import { DeviceError, UnsupportedFeatureError } from "./errors";
 import {
   BotLLMSearchResponseData,
   BotLLMTextData,
@@ -21,7 +21,7 @@ import {
   RTVIMessage,
   TranscriptData,
   UICommandData,
-  UITaskData,
+  UIJobGroupData,
 } from "./messages";
 
 export enum RTVIEvent {
@@ -42,9 +42,9 @@ export enum RTVIEvent {
   ServerResponse = "serverResponse",
   MessageError = "messageError",
 
-  /** UI Agent Protocol */
+  /** UI Worker Protocol */
   UICommand = "uiCommand",
-  UITask = "uiTask",
+  UIJobGroup = "uiJobGroup",
 
   /** service events */
   Metrics = "metrics",
@@ -107,6 +107,7 @@ export enum RTVIEvent {
   SpeakerUpdated = "speakerUpdated",
   DeviceError = "deviceError",
   MediaStateUpdated = "mediaStateUpdated",
+  UnsupportedFeature = "unsupportedFeature",
 }
 
 export type RTVIEvents = Partial<{
@@ -129,9 +130,9 @@ export type RTVIEvents = Partial<{
   serverResponse: (data: any) => void;
   messageError: (message: RTVIMessage) => void;
 
-  /** UI Agent Protocol */
+  /** UI Worker Protocol */
   uiCommand: (data: UICommandData) => void;
-  uiTask: (data: UITaskData) => void;
+  uiJobGroup: (data: UIJobGroupData) => void;
 
   /** service events */
   metrics: (data: PipecatMetricsData) => void;
@@ -193,6 +194,7 @@ export type RTVIEvents = Partial<{
   speakerUpdated: (speaker: MediaDeviceInfo) => void;
   deviceError: (error: DeviceError) => void;
   mediaStateUpdated: (mediaState: MediaState) => void;
+  unsupportedFeature: (error: UnsupportedFeatureError) => void;
 }>;
 
 export type RTVIEventHandler<E extends RTVIEvent> = E extends keyof RTVIEvents